Tests/LibWeb/Text/expected/HTML/media-source-cross-realm.txt

sourceBuffer uses iframe prototype: true
sourceBuffer uses top prototype: false
buffered uses iframe prototype: false
buffered uses top prototype: true
